requirements.txt
Django
djangorestframework
djangorestframework_simplejwt==4.3.0
django-filter
mock
Pillow
psycopg2-binary
drf-yasg
pyyaml
gunicorn
astroid==2.3.3
colorama==0.4.3
django-crispy-forms==1.8.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
pdf2image==1.13.1
requests==2.23.0
